Moving inner classes out of HeronServerTester
